Modified the upload_folder() code to allow for more than one background upload.
Each call to upload_folder() will return a 2-tuple of (upload_key, total_bytes). The key is then used to check the status of the upload, or to cancel the upload.

Also fixed some tests that used the old pyrax client names. I really should run the tests before committing. ;-)