Files that are exactly the chunk size can be uploaded as is
Previously, a file was uploaded as a single chunk if

  fsize < self.max_file_size

This is wrong: files that are exactly the max_file_size can also be
uploaded as a single chunk.  The fix is to replace < with <=:

  fsize < self.max_file_size

This bug is problematic when attempting to replicate pyrax's normal
chunk uploading behavior, while still using pyrax to upload chunks.