Fix ValueError when handling bulk_delete response.
When performing a bulk delete request, `ValueError: needs more than 1 value to unpack` is raised on line 1552, because it appears one of the lines in the response is a blank line with spaces in it.  Such a value for `resp_line` will, of course, pass the `if not resp_line` precondition but fail on `resp_key, val = resp_line.split(":", 1)`.