update version and RPM control variables for ongoing post 1.9.0 development
This represents the suggested RPM approach for post-release development.
We increment the version but set the RPM release tag to "0" which activates
generation of time and git-based RPM release strings.  I am assuming a next
version of 1.9.1

When 1.9.1 is ready for release, increment the RPM release string to "1"
as part of the commit forming the official release.  Then, as the first
post-release commit, repeat the changes seen in this commit for 1.9.2
(or whatever the appropriate next version will be).