Fixed pagination for monitoring. GitHub #433
Cloud Moinitoring uses a different pagination method than most of the
other services, which required a change to the manager class.
Furthermore, enabling this change would have been very kludgey given the
uneven class design in this module, so a major refactoring was made,
creating the proper manager classes for each resource.