Match requirements to spec and travis.
Looks like python-swiftclient was in the spec but not in the
install_requires.  I've also added nose to the testing requirements and
use the proper tests_require paramter (which may be dropped due to
ancient distutils requirements).

Would it be possible to use extras_require for bringing in the rackspace
items (to again match the spec)?