Fix memory hole while uploading 5G chunks
The call read(MAX_FILE_SIZE) will load the hole MAX_FILE_SIZE into
memory, which turns to be 5GB. That puts the system to a crawl.

Instead, the better way to go around this is to use streaming, copying
small chunks of memory in a loop until all bytes are transferred.

Apart from that:
- xreadlines has been deprecated for quite a while now:
  https://www.quantifiedcode.com/knowledge-base/maintainability/%60.xreadlines()%60%20deprecated/7lC2NACa
- Tox was not able to run pypy test env properly without the explicit
  dependencies on requests and nose.

Includes unitests for copy_maxlen_bytes:

- One test for the typical case; several writes, last one shorter
- EOF test; copy ends before maxlen due to EOF reached first
- Exact copies; when bufsize is a exact divisor of maxlen
- One copy; when maxlen is less than bufsize